IO is done in `h5geo` format

+++ src/h5geo/H5Seis.jl
@@ -1,5 +1,6 @@
 mutable struct H5Seis
   seis::PyCall.PyObject
-  ind::Array{Integer,1}
+  # types must be compatible with h5geo seis IO data
+  ind::Array{UInt64,1}
   data::Array{Float32,2}
 end
